Use auto-detection for Unity and Dbusmenu (#27)
This allows gir1.2-unity-5.0 to be moved to Recommends and not require
them in Kubuntu and Debian.

libunity9 was removed because it is indirectly used from gir1.2-unity-5.0
and it is up to that package to require it.

GObject was removed from import in scudcloud.py because it is not used
there at all.