Make the configuration file compatible with XDG
The X Desktop Guidelines recommend storing application configuration in
the ~/.config/ folder. Currently Scudcloud stores its
configuration in the ~/.scudcloud file.

This adds some functionality to relocate any legacy configuration to the
new, XDG-compliant location and then loads the new location in the
application settings.

There is an API change to the `ScudCloud` class that means it will not
run using the old launcher. That is, `ScudCloud` now expects
a `settings_path` argument and will raise an error without it.

Note: This version doesn't try to move the configuration from the
original location, per feedback from @raelgc and @immerr.

Fixes #71
Fixes #73
Fixes #75