Add mono-[dark|light] instructions to README
In #77 the user had to install the mono-dark and mono-light icon themes
as well in order to get the system tray icons to appear correctly. They
were using Arch. Since I made the Arch package based on the README,
I hadn't included these two icon themes in the package.

I recommend adding these to the README in case a user compiled it for
his own system and has the same issue.