Set INSTALL_DIR to be canonical path relative to 'scudcloud' script
INSTALL_DIR was originally hardcoded to be '/opt/scudcloud'.  This
is inflexible, especially when you want to run scudcloud directly
out of the dev tree.  This change fixes that by making the dynamically
setting the INSTALL_DIR path to be the current directory that the
scudcloud main script lives.  This gets us the original behavior plus
the ability to run scudcloud out of the dev tree.  It also works with
symlinks.

Tested on Fedora 22 with both normal installation and out of dev tree.