Allow the menubar to be hidden
On systems without unity, the menubar might look very ugly compared to
the rest of the app, so we provide the option to hide it.

We also need to register all menubar actions with shortcuts on the
mainmenu, because actions of hidden widgets do not trigger.

In case someone does not remember the shortcut, the systray also allows
toggling of the menubar.