Add --urgent-hint CLI option for window manager notifications
On X11, some window managers use the URGENCY bit in WM_HINTS to notify
the user when a window needs attention.  However, not all window
manager behave this way.

This patch adds a CLI option to scudcloud to enable setting the
URGENCY bit whenever scudcloud sends a notification:

  --urgent-hint URGENT_HINT
                        set window manager URGENT hint( default: False)

Users who want to use the URGENT hint for notifications can launch
scudcloud with the --urgent-hint argument.

Testing:

I verified this is working as expected on Ubuntu 16.04 with the xmonad
window manager.

Closes: #401
Signed-off-by: Curt Brune <curt@brune.net>